# TraceQL performance improvement: dynamic reordering of binop branches